ajout tests delete movie & user

+++ src/controllers/movieControllers.js
@@ -117,7 +117,7 @@ const postMovie = (req, res) => {
 
 const deleteMovie = (req, res) => {
   const id = parseInt(req.params.id);
-
+if(id) {
   database
     .query("delete from movies where id = ?", [id])
     .then(([result]) => {
@@ -127,11 +127,14 @@ const deleteMovie = (req, res) => {
         res.sendStatus(204);
       }
     })
+  } else {
+    database
     .catch((err) => {
       console.error(err);
-      res.sendStatus(500);
+      res.sendStatus(404);
     });
-};
+  }
+}
 
 module.exports = {
   getMovies,
